switched email_id to bigint.

+++ database/migrations/2016_06_29_110749_create_email_client_assignments_table.php
@@ -16,7 +16,7 @@ class CreateEmailClientAssignmentsTable extends Migration
     public function up()
     {
         Schema::create('email_client_assignments', function (Blueprint $table) {
-            $table->integer( 'email_id' )->unsigned();
+            $table->bigInteger( 'email_id' )->unsigned();
             $table->integer( 'client_id' )->unsigned();
             $table->timestamps();
